dev/requirements.txt
django
psycopg2-binary
pytest-django
factory-boy
